Accept directly specified resource directory.
Reasoning:
Cases like discord_arch_electron or discord-canary-electron-bin on the
AUR. These place the contents of what should be the resources directory
into /usr/lib/discord and /usr/lib/discord-canary respectively, causing
you to have to make modifications to the installation directory in order
for the installer to inject properly (moving everything in the
installation directory into a newly created 'resources' directory,
running the installer, and moving everything back). This change lets you
specify the 'resources' directory directly, so that modifications are
not necessary.